+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## [WIP]
 [WIP]: https://git.smile.fr/dirtech/gdpr-dump/compare/1.0.0-beta3...master
 
+- Allow date format in dump output file name
+- Fix dump settings in schema.json
 - Remove unused `optional` parameter in ConverterFactory
 
 ## [1.0.0-beta3] - 2019-07-01